requirements.txt
torch>=1.12.0
torchvision>=0.13.0
matplotlib
scikit-learn
pandas
Pillow
tqdm
